Fluidsynth settings

Get available settings and their types. See fluidsynth docs for more information on the available options.

Usage

fluidsynth_setting_list()

fluidsynth_setting_options(setting)

fluidsynth_setting_default(setting)

libfluidsynth_version()

Arguments

setting

string with one of the options listed in fluidsynth_setting_list(), see examples.

Value

a list with available options

References

FluidSynth Settings Reference

See also

Examples

# List available settings:
fluidsynth_setting_list()
#>                                 name    type
#> 1                  audio.alsa.device  string
#> 2                       audio.driver  string
#> 3                  audio.file.endian  string
#> 4                  audio.file.format  string
#> 5                    audio.file.name  string
#> 6                    audio.file.type  string
#> 7             audio.jack.autoconnect integer
#> 8                      audio.jack.id  string
#> 9                   audio.jack.multi integer
#> 10                 audio.jack.server  string
#> 11                  audio.oss.device  string
#> 12                 audio.period-size integer
#> 13                     audio.periods integer
#> 14   audio.pulseaudio.adjust-latency integer
#> 15           audio.pulseaudio.device  string
#> 16       audio.pulseaudio.media-role  string
#> 17           audio.pulseaudio.server  string
#> 18               audio.realtime-prio integer
#> 19               audio.sample-format  string
#> 20                 audio.sdl2.device  string
#> 21                  midi.alsa.device  string
#> 22              midi.alsa_seq.device  string
#> 23                  midi.alsa_seq.id  string
#> 24                  midi.autoconnect integer
#> 25                       midi.driver  string
#> 26                      midi.jack.id  string
#> 27                  midi.jack.server  string
#> 28                   midi.oss.device  string
#> 29                     midi.portname  string
#> 30                midi.realtime-prio integer
#> 31                player.reset-synth integer
#> 32              player.timing-source  string
#> 33                        shell.port integer
#> 34                      shell.prompt  string
#> 35              synth.audio-channels integer
#> 36                synth.audio-groups integer
#> 37               synth.chorus.active integer
#> 38                synth.chorus.depth  double
#> 39                synth.chorus.level  double
#> 40                   synth.chorus.nr integer
#> 41                synth.chorus.speed  double
#> 42                   synth.cpu-cores integer
#> 43           synth.default-soundfont  string
#> 44                   synth.device-id integer
#> 45      synth.dynamic-sample-loading integer
#> 46            synth.effects-channels integer
#> 47              synth.effects-groups integer
#> 48                        synth.gain  double
#> 49               synth.ladspa.active integer
#> 50                 synth.lock-memory integer
#> 51            synth.midi-bank-select  string
#> 52               synth.midi-channels integer
#> 53             synth.min-note-length integer
#> 54                synth.overflow.age  double
#> 55          synth.overflow.important  double
#> 56 synth.overflow.important-channels  string
#> 57         synth.overflow.percussion  double
#> 58           synth.overflow.released  double
#> 59          synth.overflow.sustained  double
#> 60             synth.overflow.volume  double
#> 61                   synth.polyphony integer
#> 62               synth.reverb.active integer
#> 63                 synth.reverb.damp  double
#> 64                synth.reverb.level  double
#> 65            synth.reverb.room-size  double
#> 66                synth.reverb.width  double
#> 67                 synth.sample-rate  double
#> 68              synth.threadsafe-api integer
#> 69                     synth.verbose integer
fluidsynth_setting_options('audio.driver')
#> [1] "alsa"       "file"       "jack"       "oss"        "pulseaudio"
#> [6] "sdl2"      
fluidsynth_setting_default('synth.sample-rate')
#> [1] 44100